+++ lib/vump/modules/component_json.rb
@@ -1,0 +1,9 @@
+require 'vump/modules/package_json'
+
+module Vump
+  class BowerJson < PackageJson
+    def filename
+      'component.json'
+    end
+  end
+end