+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Vump badge
  - Bower module
+ - Component module
 
 ## [2.0.0] - 2019-03-14
 ### Added